+++ c3smembership/accountants_views.py
@@ -320,13 +320,15 @@ def delete_entry(request):
     This view lets accountants delete entries (doublettes)
     """
 
-    deletion_confirmed = False
-    if 'deletion_confirmed' in request.params:
-        deletion_confirmed = (request.params['deletion_confirmed'] == '1')
+    deletion_confirmed = (request.params.get('deletion_confirmed', '0') == '1')
+    redirection_view = request.params.get('redirect', 'dashboard_only')
+    log.info('redirect to: ' + str(redirection_view))
 
     if deletion_confirmed:
         memberid = request.matchdict['memberid']
         _member = C3sMember.get_by_id(memberid)
+        member_lastname = _member.lastname
+        member_firstname = _member.firstname
 
         C3sMember.delete_by_id(_member.id)
         log.info(
@@ -340,17 +342,19 @@ def delete_entry(request):
         request.session.flash(_message, 'messages')
         return HTTPFound(
             request.route_url(
-                'dashboard_only',
-                _query={'message': 'Member with id {0} was deleted.'.format(
-                        memberid)}
-            ) + '#member_' + str(_member.id)
+                redirection_view,
+                _query={'message': u'Member with id {0} \"{1}, {2}\" was deleted.'.format(
+                        memberid,
+                        member_lastname,
+                        member_firstname)}
+            ) + '#member_' + str(memberid)
         )
     else:
         return HTTPFound(
             request.route_url(
-                'dashboard_only',
+                redirection_view,
                 _query={'message': 'Deleting the member was not confirmed' + \
-                    ' and therefore nothing was deleted.'}
+                    ' and therefore nothing has been deleted.'}
             )
         )
 
@@ -389,6 +393,23 @@ def switch_pay(request):
             number=dashboard_page, order=order, orderby=order_by
         ) + '#member_' + str(_member.id)
     )
+
+
+@view_config(renderer='json',
+             permission='manage',
+             route_name='get_member')
+def get_member(request):
+    memberid = request.matchdict['memberid']
+    member = C3sMember.get_by_id(memberid)
+    if member is None:
+        return {}
+    else:
+        return {
+            'id': member.id,
+            'firstname': member.firstname,
+            'lastname': member.lastname
+        }
+    return None
 
 
 @view_config(renderer='templates/detail.pt',