+++ CHANGES.md
@@ -20,6 +20,11 @@ Next Release
 
 - Added license texts section to documentation.
 
+- Added list of contributors.
+
+- Implemented redirect for member deletion based on route name.
+
+
 
 1.10.1
 ======